Joystick axis events were not being checked in the ticcmd building code, it must have been removed accidentally while merging in my code changes.
Fixed a joystick povHat issue.

Made ctl-look-speed and ctl-turn-speed (float).

+++ doomsday/Include/jDoom/d_config.h
@@ -52,8 +52,8 @@ typedef struct jdoom_config_s {
     int             joyaxis[8];
     int             jlookDeltaMode;
     int             lookSpring;
-    int             lookSpeed;
-    int             turnSpeed;
+    float           lookSpeed;
+    float           turnSpeed;
     byte            povLookAround;
     int             jumpEnabled;
     float           jumpPower;
