+++ doomsday/Include/jHeretic/h_config.h
@@ -51,8 +51,8 @@ typedef struct jheretic_config_s {
     int             joyaxis[8];
     int             jlookDeltaMode;
     int             lookSpring;
-    int             lookSpeed;
-    int             turnSpeed;
+    float           lookSpeed;
+    float           turnSpeed;
     byte            povLookAround;
     int             jumpEnabled;
     float           jumpPower;
